+++ sdk/keyvault/azure-mgmt-keyvault/azure/mgmt/keyvault/v2016_10_01/aio/operations/_vaults_operations.py
@@ -48,7 +48,7 @@ class VaultsOperations:
         resource_group_name: str,
         vault_name: str,
         parameters: "_models.VaultCreateOrUpdateParameters",
-        **kwargs
+        **kwargs: Any
     ) -> "_models.Vault":
         """Create or update a key vault in the specified subscription.
 
@@ -118,7 +118,7 @@ class VaultsOperations:
         resource_group_name: str,
         vault_name: str,
         parameters: "_models.VaultPatchParameters",
-        **kwargs
+        **kwargs: Any
     ) -> "_models.Vault":
         """Update a key vault in the specified subscription.
 
@@ -187,7 +187,7 @@ class VaultsOperations:
         self,
         resource_group_name: str,
         vault_name: str,
-        **kwargs
+        **kwargs: Any
     ) -> None:
         """Deletes the specified Azure key vault.
 
@@ -240,7 +240,7 @@ class VaultsOperations:
         self,
         resource_group_name: str,
         vault_name: str,
-        **kwargs
+        **kwargs: Any
     ) -> "_models.Vault":
         """Gets the specified Azure key vault.
 
@@ -300,7 +300,7 @@ class VaultsOperations:
         vault_name: str,
         operation_kind: Union[str, "_models.AccessPolicyUpdateKind"],
         parameters: "_models.VaultAccessPolicyParameters",
-        **kwargs
+        **kwargs: Any
     ) -> "_models.VaultAccessPolicyParameters":
         """Update access policies in a key vault in the specified subscription.
 
@@ -372,7 +372,7 @@ class VaultsOperations:
         self,
         resource_group_name: str,
         top: Optional[int] = None,
-        **kwargs
+        **kwargs: Any
     ) -> AsyncIterable["_models.VaultListResult"]:
         """The List operation gets information about the vaults associated with the subscription and
         within the specified resource group.
@@ -447,7 +447,7 @@ class VaultsOperations:
     def list_by_subscription(
         self,
         top: Optional[int] = None,
-        **kwargs
+        **kwargs: Any
     ) -> AsyncIterable["_models.VaultListResult"]:
         """The List operation gets information about the vaults associated with the subscription.
 
@@ -517,7 +517,7 @@ class VaultsOperations:
 
     def list_deleted(
         self,
-        **kwargs
+        **kwargs: Any
     ) -> AsyncIterable["_models.DeletedVaultListResult"]:
         """Gets information about the deleted vaults in a subscription.
 
@@ -585,7 +585,7 @@ class VaultsOperations:
         self,
         vault_name: str,
         location: str,
-        **kwargs
+        **kwargs: Any
     ) -> "_models.DeletedVault":
         """Gets the deleted Azure key vault.
 
@@ -643,7 +643,7 @@ class VaultsOperations:
         self,
         vault_name: str,
         location: str,
-        **kwargs
+        **kwargs: Any
     ) -> None:
         cls = kwargs.pop('cls', None)  # type: ClsType[None]
         error_map = {
@@ -685,7 +685,7 @@ class VaultsOperations:
         self,
         vault_name: str,
         location: str,
-        **kwargs
+        **kwargs: Any
     ) -> AsyncLROPoller[None]:
         """Permanently deletes the specified vault. aka Purges the deleted Azure key vault.
 
@@ -695,8 +695,8 @@ class VaultsOperations:
         :type location: str
         :keyword callable cls: A custom type or function that will be passed the direct response
         :keyword str continuation_token: A continuation token to restart a poller from a saved state.
-        :keyword polling: Pass in True if you'd like the AsyncARMPolling polling method,
-         False for no polling, or your own initialized polling object for a personal polling strategy.
+        :keyword polling: By default, your polling method will be AsyncARMPolling.
+         Pass in False for this operation to not poll, or pass in your own initialized polling object for a personal polling strategy.
         :paramtype polling: bool or ~azure.core.polling.AsyncPollingMethod
         :keyword int polling_interval: Default waiting time between two polls for LRO operations if no Retry-After header is present.
         :return: An instance of AsyncLROPoller that returns either None or the result of cls(response)
@@ -748,7 +748,7 @@ class VaultsOperations:
     def list(
         self,
         top: Optional[int] = None,
-        **kwargs
+        **kwargs: Any
     ) -> AsyncIterable["_models.ResourceListResult"]:
         """The List operation gets information about the vaults associated with the subscription.
 
@@ -821,7 +821,7 @@ class VaultsOperations:
     async def check_name_availability(
         self,
         vault_name: "_models.VaultCheckNameAvailabilityParameters",
-        **kwargs
+        **kwargs: Any
     ) -> "_models.CheckNameAvailabilityResult":
         """Checks that the vault name is valid and is not already in use.